remove freezed & sanitize attendances/overview
