+++ lib/src/application/attendances/overview/event.dart
@@ -1,6 +1,9 @@
 part of 'bloc.dart';
 
-@freezed
-class AttendancesOverviewEvent with _$AttendancesOverviewEvent {
-  const factory AttendancesOverviewEvent.started() = _Started;
+sealed class AttendancesOverviewEvent {
+  const AttendancesOverviewEvent();
+}
+
+class AttendancesOverviewStarted extends AttendancesOverviewEvent {
+  const AttendancesOverviewStarted();
 }
