+++ lib/src/application/attendances/overview/state.dart
@@ -1,10 +1,14 @@
 part of 'bloc.dart';
 
-@freezed
-class AttendancesOverviewState with _$AttendancesOverviewState {
-  const factory AttendancesOverviewState.loading() = AttendancesOverviewLoading;
+sealed class AttendancesOverviewState {
+  const AttendancesOverviewState();
+}
 
-  const factory AttendancesOverviewState.success({
-    required List<Attendance> attendances,
-  }) = AttendancesOverviewSuccess;
+class AttendancesOverviewLoadInProgress extends AttendancesOverviewState {
+  const AttendancesOverviewLoadInProgress();
+}
+
+class AttendancesOverviewLoadSuccess extends AttendancesOverviewState {
+  const AttendancesOverviewLoadSuccess(this.attendances);
+  final List<Attendance> attendances;
 }
