--- lib/src/presentation/attendances/overview/widgets/attendances_overview_view.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-import 'body/body.dart';
-
-class AttendancesOverviewView extends StatelessWidget {
-  const AttendancesOverviewView({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const AttendancesOverviewBody();
-  }
-}
